+++ dart_ping/lib/dart_ping.dart
@@ -5,5 +5,7 @@ export 'package:dart_ping/src/models/ping_error.dart';
 export 'package:dart_ping/src/models/ping_response.dart';
 export 'package:dart_ping/src/models/ping_summary.dart';
 export 'package:dart_ping/src/ip_version.dart';
+export 'package:dart_ping/src/address_family.dart'
+    show ipLiteralFamily, validateAddressFamily;
 export 'package:dart_ping/src/ping_interface.dart';
 export 'package:dart_ping/src/models/ping_parser.dart';